Remove compiler warnings.

+++ src/plugins/id3/id3.c
@@ -648,7 +648,7 @@ _parse_id3v1(int fd, struct id3_info *info, lms_charset_conv_t *cs_conv)
 static void *
 _match(struct plugin *p, const char *path, int len, int base)
 {
-    int i;
+    long i;
 
     i = lms_which_extension(path, len, _exts, LMS_ARRAY_SIZE(_exts));
     if (i < 0)
@@ -717,8 +717,8 @@ _parse(struct plugin *plugin, struct lms_context *ctxt, const struct lms_file_in
     }
 
     if (!info.title.str) {
-        int ext_idx;
-        ext_idx = ((int)match) - 1;
+        long ext_idx;
+        ext_idx = ((long)match) - 1;
         info.title.len = finfo->path_len - finfo->base - _exts[ext_idx].len;
         info.title.str = malloc((info.title.len + 1) * sizeof(char));
         memcpy(info.title.str, finfo->path + finfo->base, info.title.len);
