+++ src/plugins/ogg/ogg.c
@@ -240,7 +240,7 @@ struct plugin {
 static void *
 _match(struct plugin *p, const char *path, int len, int base)
 {
-    int i;
+    long i;
 
     i = lms_which_extension(path, len, _exts, LMS_ARRAY_SIZE(_exts));
     if (i < 0)
@@ -260,9 +260,9 @@ _parse(struct plugin *plugin, struct lms_context *ctxt, const struct lms_file_in
       goto done;
 
     if (!info.title.str) {
-      int ext_idx;
+      long ext_idx;
 
-      ext_idx = ((int)match) - 1;
+      ext_idx = ((long)match) - 1;
       info.title.len = finfo->path_len - finfo->base - _exts[ext_idx].len;
       info.title.str = (char *)malloc((info.title.len + 1) * sizeof(char));
       memcpy(info.title.str, finfo->path + finfo->base, info.title.len);
